全警告を抹殺

+++ DebugCamera.cpp
@@ -4,20 +4,20 @@
 void DebugCamera::Update()
 {
 	if (RInput::GetKey(DIK_W)) {
-		viewProjection.eye.x += -moveSpeed * cos(Util::PI / 180 * angle.x);
-		viewProjection.eye.z += moveSpeed * sin(Util::PI / 180 * angle.x);
+		viewProjection.eye.x += -moveSpeed * cosf(Util::PI / 180 * angle.x);
+		viewProjection.eye.z += moveSpeed * sinf(Util::PI / 180 * angle.x);
 	}
 	if (RInput::GetKey(DIK_S)) {
-		viewProjection.eye.x += -moveSpeed * cos(Util::PI / 180 * (angle.x + 180));
-		viewProjection.eye.z += moveSpeed * sin(Util::PI / 180 * (angle.x + 180));
+		viewProjection.eye.x += -moveSpeed * cosf(Util::PI / 180 * (angle.x + 180));
+		viewProjection.eye.z += moveSpeed * sinf(Util::PI / 180 * (angle.x + 180));
 	}
 	if (RInput::GetKey(DIK_A)) {
-		viewProjection.eye.x += -moveSpeed * cos(Util::PI / 180 * (angle.x - 90));
-		viewProjection.eye.z += moveSpeed * sin(Util::PI / 180 * (angle.x - 90));
+		viewProjection.eye.x += -moveSpeed * cosf(Util::PI / 180 * (angle.x - 90));
+		viewProjection.eye.z += moveSpeed * sinf(Util::PI / 180 * (angle.x - 90));
 	}
 	if (RInput::GetKey(DIK_D)) {
-		viewProjection.eye.x += -moveSpeed * cos(Util::PI / 180 * (angle.x + 90));
-		viewProjection.eye.z += moveSpeed * sin(Util::PI / 180 * (angle.x + 90));
+		viewProjection.eye.x += -moveSpeed * cosf(Util::PI / 180 * (angle.x + 90));
+		viewProjection.eye.z += moveSpeed * sinf(Util::PI / 180 * (angle.x + 90));
 	}
 	if (RInput::GetKey(DIK_SPACE)) {
 		viewProjection.eye.y += moveSpeed;
@@ -43,9 +43,9 @@ void DebugCamera::Update()
 		angle.y = -89.9f;
 	}
 
-	viewProjection.target.x = viewProjection.eye.x + -100 * cos(Util::PI / 180 * angle.x) * cos(Util::PI / 180 * angle.y);
-	viewProjection.target.y = viewProjection.eye.y + 100 * sin(Util::PI / 180 * angle.y);
-	viewProjection.target.z = viewProjection.eye.z + 100 * sin(Util::PI / 180 * angle.x) * cos(Util::PI / 180 * angle.y);
+	viewProjection.target.x = viewProjection.eye.x + -100 * cosf(Util::PI / 180 * angle.x) * cosf(Util::PI / 180 * angle.y);
+	viewProjection.target.y = viewProjection.eye.y + 100 * sinf(Util::PI / 180 * angle.y);
+	viewProjection.target.z = viewProjection.eye.z + 100 * sinf(Util::PI / 180 * angle.x) * cosf(Util::PI / 180 * angle.y);
 
 	viewProjection.UpdateMatrix();
 }
